refactor: rename lumberjack.log.ts
