+++ libs/ngworker/lumberjack/src/lib/configuration/lumberjack-log-format-function.ts
@@ -1,3 +1,3 @@
-import { LumberjackLog } from '../logs/lumberjack-log';
+import { LumberjackLog } from '../logs/lumberjack.log';
 
 export type LumberjackLogFormatFunction = (logEntry: LumberjackLog) => string;
